api:fix - Error when not exists analysis in database
When not exists analysis in database, I founded this problem.
And for solution it was necessary change this logic for check if
not exists hashes deprecated. Other problem resolved here was
about the responsability of the methods, them I change all code
with database logic to repositories layer and stay only business logic
into controllers layer.

Signed-off-by: Wilian Gabriel <wilian.silva@zup.com.br>